Fix xmlrpc_error:number handling

+++ main/php_globals.h
@@ -125,7 +125,7 @@ struct _php_core_globals {
 	zend_bool html_errors;
 	zend_bool xmlrpc_errors;
 
-	short xmlrpc_error_number;
+	long xmlrpc_error_number;
 
 
 	zend_bool modules_activated;
